arch/x86: Fix debug delay overflow and optimize compat syscalls
* arch_debug.cpp: Change delay type to uint64 to prevent overflow when calculating TSC delay for large durations (> 1.4s at 3GHz).
* syscalls.cpp (x86_64): Implement vendor check in x86_compat_initialize_syscall to use the SYSCALL instruction stub for AMD/Hygon CPUs and SYSENTER for Intel. AMD processors do not support SYSENTER in long mode (it raises #UD), so 32-bit applications must use SYSCALL. Resolves TODO: "check AMD for sysenter".

+++ src/system/kernel/arch/x86/arch_debug.cpp
@@ -1391,7 +1391,7 @@ debug_snooze_tpause(uint32 delay)
 void
 arch_debug_snooze(bigtime_t duration)
 {
-	uint32 delay = (duration * sDebugSnoozeConversionFactor) / 1000;
+	uint64 delay = (duration * sDebugSnoozeConversionFactor) / 1000;
 	if (delay == 0)
 		delay = 1;
